refactor: [DD-1923] Add export for ResizeObserverComponent
