+++ packages/components/src/ResizeObserverComponent/ResizeObserverComponent.tsx
