+++ packages/components/src/ResizeObserverComponent/ResizeObserverPolyfill.ts
