+++ packages/components/src/index.ts
@@ -30,6 +30,7 @@ export { HiddenSvgDefs } from './HiddenSvgDefs/HiddenSvgDefs';
 export { DialogContent } from './DialogContent/DialogContent';
 export { AdaptiveGrid as Grid } from './AdaptiveGrid/AdaptiveGrid';
 export { CurrencyIcon } from './CurrencyIcon/CurrencyIcon';
+export { ResizeObserverComponent } from './ResizeObserverComponent/ResizeObserverComponent';
 
 export * from './inputs';
 export * from './icons';
